requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
bcrypt==4.1.2
python-slugify==8.0.1
